susefirewall2-to-firewalld: Stop guessing firewall service from port/proto
Some services share the same port/proto combinations so trying to guess
which one we want is error-prone. The only reason to do that is to make
the output more readable and use a single service instead of the
individual ports it needs. However, this script only aims to provide
some basic functionality on best-effort basis and as such it's best
to simply add the ports themselves based on the iptables output.

Fixes: #6