Fix `ERROR: INVALID_PORT` when `--sport` iptables rules are present
When there are certain iptables rules containing only an `--sport` but
no `--dport` then the execution of the script fails. See openSUSE bug
[1].

This change only adds robustness by skipping over such occurences. It
doesn't actually process the entries in question. This means the result
of the migration probably was and still will be incomplete in some
constellations.

[1]: https://bugzilla.suse.com/show_bug.cgi?id=1170461